connect4/requirements.txt
Flask
Flask-SocketIO
paho-mqtt
flask_cors